First pass at unifying product item views
