+++ WooCommerce/src/main/res/layout/product_list_item.xml
@@ -1,93 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/linearLayout"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@drawable/list_primary_selector">
 
-    <FrameLayout
-        android:id="@+id/productImageFrame"
-        android:layout_width="@dimen/image_minor_100"
-        android:layout_height="@dimen/image_minor_100"
-        android:layout_margin="@dimen/major_100"
-        android:background="@drawable/picture_frame"
-        android:padding="1dp"
-        app:layout_constraintVertical_bias="0.50"
-        app:layout_constraintBottom_toTopOf="@+id/divider"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageView
-            android:id="@+id/productImage"
-            android:layout_width="@dimen/image_minor_100"
-            android:layout_height="@dimen/image_minor_100"
-            android:layout_gravity="center"
-            android:background="@drawable/picture_corners"
-            android:importantForAccessibility="no"
-            android:scaleType="centerCrop"
-            tools:src="@drawable/ic_product" />
-    </FrameLayout>
-
-    <!--
-        LinearLayout is necessary to ensure the contents will always center properly in the
-        view, even when the productStockAndStatus is hidden, or if productName takes up
-        two lines.
-    -->
-    <LinearLayout
-        android:id="@+id/productInfoContainer"
-        android:layout_width="0dp"
+    <com.woocommerce.android.ui.products.ProductItemView
+        android:id="@+id/productItemView"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="@dimen/major_100"
-        android:orientation="vertical"
-        app:layout_constrainedWidth="true"
-        app:layout_constraintBottom_toTopOf="@+id/divider"
-        app:layout_constraintEnd_toStartOf="@+id/product_btnDelete"
-        app:layout_constraintStart_toEndOf="@+id/productImageFrame"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.50">
-
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/productName"
-            style="@style/Woo.ListItem.Title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/minor_00"
-            android:ellipsize="end"
-            android:includeFontPadding="false"
-            android:maxLines="2"
-            tools:text="Awesome Sauce and all that jazz is all I gotta tell you how do we go " />
-
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/productStockAndStatus"
-            style="@style/Woo.ListItem.Body"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/minor_00"
-            tools:text="Out of stock this is a really long version to see what happens when we move o" />
-
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/productSku"
-            style="@style/Woo.ListItem.Body"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/minor_00"
-            tools:text="SKU" />
-    </LinearLayout>
-
-    <ImageButton
-        android:id="@+id/product_btnDelete"
-        style="@style/Woo.ImageButton.Close"
-        android:contentDescription="@string/grouped_product_btn_delete"
-        android:scaleType="centerInside"
-        android:layout_margin="@dimen/minor_100"
-        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:visibility="gone"
-        tools:visibility="visible"/>
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <View
         android:id="@+id/divider"
@@ -97,6 +22,6 @@
         android:layout_marginBottom="@dimen/minor_00"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@+id/productInfoContainer"/>
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
